+++ SRC/Injector/Public/Extensions/IServiceProviderBasicExtensions.cs
@@ -1,0 +1,9 @@
+﻿/********************************************************************************
+* IServiceProviderBasicExtensions.cs                                            *
+*                                                                               *
+* Author: Denes Solti                                                           *
+********************************************************************************/
+using System.Runtime.CompilerServices;
+using Solti.Utils.DI.Interfaces;
+
+[assembly: TypeForwardedTo(typeof(IServiceProviderBasicExtensions))]