+++ SRC/Interfaces/Extensions/IServiceProviderBasicExtensions.cs
@@ -1,27 +1,26 @@
-﻿/********************************************************************************
-* IServiceProviderExtensions.cs                                                 *
+/********************************************************************************
+* IServiceProviderBasicExtensions.cs                                            *
 *                                                                               *
 * Author: Denes Solti                                                           *
 ********************************************************************************/
 using System;
 
-namespace Solti.Utils.DI
+namespace Solti.Utils.DI.Interfaces
 {
-    using Internals;
-
     /// <summary>
-    /// Defines several handy extensions for the <see cref="IServiceProvider"/> interface.
+    /// Defines basic extensions for the <see cref="IServiceProvider"/> interface.
     /// </summary>
-    public static class IServiceProviderExtensions
+    public static class IServiceProviderBasicExtensions
     {
         /// <summary>
         /// Gets the service instance associated with the given interface.
         /// </summary>
         public static TInterface? GetService<TInterface>(this IServiceProvider self) where TInterface : class
         {
-            Ensure.Parameter.IsNotNull(self, nameof(self));
+            if (self == null)
+                throw new ArgumentNullException(nameof(self));
 
-            return (TInterface) self.GetService(typeof(TInterface));
+            return (TInterface?) self.GetService(typeof(TInterface));
         }
     }
 }